Isolate NMRPCA classes to the separate module. Add recomputing block to analyze.py for testing.
